+++ satella/coding/structures/dictionaries.py
@@ -1,5 +1,6 @@
 import collections.abc
 import copy
+import time
 import typing as tp
 
 from satella.coding.recast_exceptions import rethrow_as