+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 * added `postcondition`
 * added __slots__ to `ReprableMixin` and `Multirun`
 * added `read_in_file` and `write_to_file`
+* optimized `TimeBasedHeap`
+* added `TimeBasedSetHeap`